testing directly against electron with tap
